Move all k6 browser module definitions into one
There's an issue with importing the type definitions for the k6
browser module where the actual definitions aren't displaying but
the types can be imported whereby the IDE recognises the types.

We've found that by moving them into a single file helps resolves this
issue.
