--- types/k6/experimental/browser/worker.d.ts
@@ -1,13 +1,0 @@
-import { PageFunction } from '.';
-import { JSHandle } from './js_handle';
-
-/**
- * The Worker class represents a [WebWorker](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API).
- */
-export class Worker {
-    /**
-     * Get the URL of the web worker.
-     * @return The URL of the web worker.
-     */
-    url(): string;
-}
